Refactored.

+++ src/main/java/main/dao/Dao.java
@@ -9,5 +9,5 @@ public interface Dao<T> {
     List<T> getAll() throws SQLException;
     void save(T t) throws SQLException;
     void update(T t) throws SQLException;
-    Optional<T> getById(int id) throws SQLException;
+    Optional<T> getById(long id) throws SQLException;
 }
